+++ src/main/java/project/finaltoyproject/domain/order/dto/OrderResponseDto.java
@@ -5,17 +5,19 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import project.finaltoyproject.domain.order.OrderState;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class OrderResponseDto {
     private Long orderId;
     private OrderState orderState;
-    private Integer totalOrderPrice;
+    private double totalOrderPrice;
     private String username;
     private String itemName;
     private int remainQuantity;
 
-    public OrderResponseDto(Long orderId, OrderState orderState, Integer totalOrderPrice, String username, String itemName, int remainQuantity) {
+    public OrderResponseDto(Long orderId, OrderState orderState, double totalOrderPrice, String username, String itemName, int remainQuantity) {
         this.orderId = orderId;
         this.orderState = orderState;
         this.totalOrderPrice = totalOrderPrice;
